+++ ising/ext_sources/search.cu
@@ -39,7 +39,7 @@ void find_lowest(
   cudaMalloc((void **) &d_low_en, sizeof(T) * num_st * 2);
   cudaMalloc((void **) &d_low_st, sizeof(long int) * num_st *2);
   
-  for(int m=0; m < pow(2, N - chunk_exp); m++) {
+  for(long int m=0; m < pow(2, N - chunk_exp); m++) {
     callback(m, user_data);
     idx = m * chunk_size;
     search<<<grid_size, block_size>>>(d_Jh, N, chunk_size, d_energies, d_states, idx);
